After a day of testing...
This version attempts once again to calculate the position from the accelerometer; it follows the advice given us online. The accelerometer predictably returns a certain offset when tilted a certain amount; this version calculates true angles and subtracts that offset. Please try this version. It prints the calculated displacements, in X and in Y, to the serial monitor.